+++ lib/src/cook/OS.cpp
@@ -5,12 +5,16 @@ namespace cook {
 
 OS get_os()
 {
+#if BOOST_OS_LINUX
+    return OS::Linux;
+#endif
+
 #if BOOST_OS_WINDOWS
     return OS::Windows;
 #endif
 
-#if BOOST_OS_LINUX
-    return OS::Linux;
+#if BOOST_OS_MACOS
+    return OS::MacOS;
 #endif
 }
 
@@ -19,8 +23,9 @@ std::ostream & operator<<(std::ostream & str, OS os)
     switch(os)
     {
 #define L_CASE(NAME) case OS::NAME: return str << #NAME
-        L_CASE(Windows);
         L_CASE(Linux);
+        L_CASE(Windows);
+        L_CASE(MacOS);
 #undef L_CASE
     }
 }