+++ changelog.md
@@ -29,6 +29,7 @@ Proof-of-concept version.
 
 ## 1.1.1 (open)
 
+* Added `-C void` chef to speed-up generation of naft output when only the recipe names are required.
 * [wip] Support for gcc, clang and msvc toolchains
 * [wip] Support for windows and osx